// src/app/shared/types/api-types/bidt-proc-req-item-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtDfItemDTO } from './bidt-df-item-dto.interface';
import { BidtProcReqItemIdDTO } from './bidt-proc-req-item-id-dto.interface';
import { BidtProcurementRequestDTO } from './bidt-procurement-request-dto.interface';
import { BidtStockMvtDTO } from './bidt-stock-mvt-dto.interface';
import { BidtTransferOrderDTO } from './bidt-transfer-order-dto.interface';
import { BirePnDTO } from './bire-pn-dto.interface';

export interface BidtProcReqItemDTO extends BidtProcReqItemIdDTO {
  bidtDfItem?: string;
  bidtDfItems?: BidtDfItemDTO[];
  bidtTorItem?: string;
  bidtTransferOrders?: BidtTransferOrderDTO[];
  priCode?: string; // 50
  priStatus?: string; // 50
  requestedQuantity?: string;
  statusDate?: Date;
  statusState?: string; // 50
  statusUser?: string; // 100
  birePn?: BirePnDTO;
  birePnPnCode?: string;
  mandant?: string;
  birePnMandant?: string;
  bidtDfItemId?: number;
  bidtProcurementRequestId?: number;
  bidtProcurementRequest?: BidtProcurementRequestDTO;
  bidtTorItemId?: number;
  workshopEntries?: BidtStockMvtDTO[];
  bidtWarehouseId?: number;
}